implementing suggested changes from PR discussion
