+++ .devcontainer/devcontainer.json
@@ -1,0 +1,6 @@
+{
+	"image": "mcr.microsoft.com/devcontainers/universal:2",
+	"features": {
+		"ghcr.io/devcontainers/features/rust:1": {}
+	}
+}
