+++ examples/simple-protected-endpoint/main.rs
@@ -19,5 +19,5 @@ async fn handler(guard: FirebaseToken) -> String {
     // Including the FirebaseToken guard is enough
     // the handler will run only if the token is valid.
     // The request guard won't work if FirebaseAuth state is not present. 
-    format!("Hello, you're logged in as user ID {}", guard.token.uid)
+    format!("Hello, you're logged in as user ID {}", guard.sub)
 }
